% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\docType{methods}
\name{back_propogation,Neural_Network,numeric,numeric,numeric-method}
\alias{back_propogation,Neural_Network,numeric,numeric,numeric-method}
\title{back prop}
\usage{
\S4method{back_propogation}{Neural_Network,numeric,numeric,numeric}(object,
  target, regularization_parameter, learning_rate)
}
\arguments{
\item{object}{is a Neural_Network}

\item{target}{is a numeric vector}

\item{regularization_parameter}{is non-negative number punishes strong connections}

\item{learning_rate}{is a positive number that controls the rate at which connections are adjusted}
}
\value{
Neural_Network
}
\description{
updates connection strengths using results of last forward prop
}

