% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\docType{methods}
\name{forward_propogation,Neural_Network,matrix-method}
\alias{forward_propogation,Neural_Network,matrix-method}
\title{f_prop}
\usage{
\S4method{forward_propogation}{Neural_Network,matrix}(object, dataset)
}
\arguments{
\item{object}{is a Neural_Network}

\item{dataset}{is a matrix not containing the target vector}
}
\value{
Neural_Network
}
\description{
... part of the training program
}

