% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\docType{methods}
\name{Get_Cost,Neural_Network,numeric-method}
\alias{Get_Cost,Neural_Network,numeric-method}
\title{cost}
\usage{
\S4method{Get_Cost}{Neural_Network,numeric}(object, target)
}
\arguments{
\item{object}{... a Neural_Network that has run forward_prop at least once}

\item{target}{... a numeric vector ... the target ...}
}
\value{
Numeric
}
\description{
get the logarithmic loss for a set of predictions
}

