% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\name{Get_LogLoss}
\alias{Get_LogLoss}
\title{log loss}
\usage{
Get_LogLoss(predictions, target)
}
\arguments{
\item{predictions}{is a numeric vector}

\item{target}{is a numeric vector}
}
\value{
Numeric
}
\description{
get log loss
}

