% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\docType{methods}
\name{Get_Number_Observations,Neural_Network-method}
\alias{Get_Number_Observations,Neural_Network-method}
\title{num observs}
\usage{
\S4method{Get_Number_Observations}{Neural_Network}(object)
}
\arguments{
\item{object}{... a Neural Network that has called fprop. ie. that has called train/predict}
}
\value{
Numeric
}
\description{
returns the number of observations that the network has processed
}

