% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\docType{methods}
\name{initialize,Neural_Network-method}
\alias{initialize,Neural_Network-method}
\title{init}
\usage{
\S4method{initialize}{Neural_Network}(.Object, number_predictors,
  hidden_layer_lengths)
}
\arguments{
\item{.Object}{... a Neural_Network object}

\item{number_predictors}{... a numeric telling how many preditors there are}

\item{hidden_layer_lengths}{... a numeric telling the number of layers and the number of neurons in each layer}
}
\value{
Neural_Network
}
\description{
initalizes a neural network capable of studying datasets with ncol = to the ncol(sample_dataset) and making predictions on such datasets
}
\details{
NN is parametrized by its connection_strength matrices
}

