% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\docType{methods}
\name{Predict,Neural_Network,data.frame-method}
\alias{Predict,Neural_Network,data.frame-method}
\title{predict stuff}
\usage{
\S4method{Predict}{Neural_Network,data.frame}(object, dataset)
}
\arguments{
\item{object}{: a neural network}

\item{dataset}{: a dataframe of features and observations}
}
\value{
Numeric
}
\description{
returns predictions
}

