% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\name{Start}
\alias{Start}
\title{start script}
\usage{
Start()
}
\description{
main function that runs the interactive script
}
\details{
takes your numerai training data and trains a neural network to your architectural specifications.
provides you with the out of sample error
offers to retrain with a new architecture or predict on a numerai tournament dataset.
Can then write the predictions to a CSV
}

