% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamiaNN.R
\docType{methods}
\name{Train,Neural_Network,data.frame,numeric,numeric,numeric-method}
\alias{Train,Neural_Network,data.frame,numeric,numeric,numeric-method}
\title{train the NN}
\usage{
\S4method{Train}{Neural_Network,data.frame,numeric,numeric,numeric}(object,
  dataset, regularization_constant, learning_rate, tolerable_error)
}
\arguments{
\item{object}{is a Neural Network}

\item{dataset}{is a data.frame, the original data frame that includes the target}

\item{regularization_constant}{is a numeric}

\item{learning_rate}{is a numeric}

\item{tolerable_error}{is a numeric, units : log loss}
}
\value{
Neural_Network
}
\description{
gets NN parameters that minimize cost on dataset using optimization_method
}

